requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cmake==4.1.3
cryptography==46.0.3
flatbuffers==25.9.23
frozenlist==1.8.0
gcloud==0.18.3
geographiclib==2.1
geopy==2.4.1
googleapis-common-protos==1.72.0
h3==4.3.1
httplib2==0.31.0
idna==3.11
jwcrypto==1.5.6
magic-filter==1.0.12
multidict==6.7.0
numpy==2.3.5
oauth2client==4.1.3
propcache==0.4.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
Pyrebase4==4.8.0
python-jwt==4.1.0
pytz==2025.2
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==0.10.1
rsa==4.9.1
six==1.17.0
timezonefinder==8.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
yarl==1.22.0
python-dotenv==1.0.0
